+++ src/psd.ts
@@ -698,6 +698,8 @@ export interface KeyDescriptorItem {
 		bottom: UnitsValue;
 		right: UnitsValue;
 	};
+	keyOriginBoxCorners?: { x: number; y: number; }[];
+	transform?: number[]; // [xx, xy, yx, yy, tx, ty]
 }
 
 export interface LayerAdditionalInfo {
@@ -964,4 +966,6 @@ export interface WriteOptions {
 	/** Forces bottom layer to be treated as layer and not background even when it's missing any transparency
 	 * 	(by default Photoshop treats bottom layer as background it it doesn't have any transparent pixels) */
 	noBackground?: boolean;
+	/** Saves document as PSB (Large Document Format) file */
+	psb?: boolean;
 }